Updated routes

+++ backend/db/migrations/02-create-spot.js
@@ -34,11 +34,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(9, 7),
+        type: Sequelize.FLOAT(9, 7),
         allowNull: false
       },
       lng: {
-        type: Sequelize.DECIMAL(10, 7),
+        type: Sequelize.FLOAT(10, 7),
         allowNull: false
       },
       name: {
